+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/envs/rl_env_cfg.py
@@ -11,11 +11,15 @@ from omni.isaac.orbit.command_generators import CommandGeneratorBaseCfg
 from omni.isaac.orbit.utils import configclass
 
 from .base_env_cfg import BaseEnvCfg
+from .ui import RLEnvWindow
 
 
 @configclass
 class RLEnvCfg(BaseEnvCfg):
     """Configuration for a reinforcement learning environment."""
+
+    # ui settings
+    ui_window_class_type: type | None = RLEnvWindow
 
     # general settings
     episode_length_s: float = MISSING